test(evm): split archived StakingRewards suite
